Páginação retirada e consultas alteradas

+++ src/main/java/com/destino/projeto_destino/dto/compra/CompraRequestDTO.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 
 public record CompraRequestDTO(
         UUID usuarioId,
-        int pacoteId,
+        long pacoteId,
         Metodo metodo,          // VISTA, PARCELADO
         Processador processador, // VISA, MASTERCARD, PIX
         int parcelas            // 1 a 12
